create useStatsFilters
